issue/10

+++ includes/class-track-history.php
@@ -3,10 +3,13 @@
  * Functionality for tracking and saving customer history.
  *
  * @package EDD User History
- * @author rzen Media, LLC
+ * @author  Easy Digital Downloads, LLC
  * @license GPL2
- * @link https://rzen.net
  */
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
 
 class EDDUH_Track_History {
 
@@ -16,40 +19,35 @@ class EDDUH_Track_History {
 	 * @since 1.5.0
 	 */
 	public function __construct() {
-
-		// Connect to WooCommerce
 		add_action( 'edduh_visited_url', array( $this, 'update_customer_history' ), 10, 3 );
 		add_action( 'edd_payment_meta', array( $this, 'save_customer_history' ) );
 
 		// Uncomment the following action to enable devmode
 		// add_action( 'get_header', array( $this, 'devmode' ) );
-
-	} /* __construct() */
+	}
 
 	/**
 	 * Get customer's tracked browsing history.
 	 *
-	 * @since 1.5.0
+	 * @since    1.5.0
+	 *
+	 * @referrer string The Referrer for ths user
 	 */
 	private function get_customer_history( $referrer = '' ) {
-
-		$user_hash = EDDUH_Cookie_Helper::get_cookie();
+		$user_hash        = EDDUH_Cookie_Helper::get_cookie();
 		$customer_history = edduh_get_page_history( $user_hash );
 
 		// If user has an established history, return that
 		if ( ! empty( $customer_history ) ) {
 			return (array) $customer_history;
 
-		// Otherwise, return an array with the original referrer
+			// Otherwise, return an array with the original referrer
 		} else {
-			$referrer = esc_url( $referrer )
-				? $referrer
-				: __( 'Direct Traffic', 'woocommerce-customer-history' );
+			$referrer = esc_url( $referrer ) ? $referrer : __( 'Direct Traffic', 'edduh' );
 
 			return array( array( 'url' => $referrer, 'time' => time() ) );
 		}
-
-	} /* get_customer_history() */
+	}
 
 	/**
 	 * Update customer's tracked browsing history.
@@ -57,26 +55,25 @@ class EDDUH_Track_History {
 	 * @since 1.5.0
 	 */
 	public function update_customer_history( $page_url = '', $timestamp = 0, $referrer = '' ) {
-
 		// Grab browsing history from the current session
-		$history = $this->get_customer_history( $referrer );
+		$history   = $this->get_customer_history( $referrer );
 		$history[] = array( 'url' => esc_url( $page_url ), 'time' => absint( $timestamp ) );
 
 		// Push the updated history to the current session
 		$user_hash = EDDUH_Cookie_Helper::get_cookie();
 		edduh_set_page_history( $user_hash, $history );
-
-	} /* update_customer_history() */
+	}
 
 	/**
 	 * Save user history as payment meta.
 	 *
 	 * @since 1.5.0
 	 *
-	 * @param array $payment_meta EDD Payment meta information.
+	 * @param  array $payment_meta EDD Payment meta information.
+	 *
+	 * @return array $payment_meta
 	 */
 	public function save_customer_history( $payment_meta = array() ) {
-
 		// Bail early if not on the purchase screen
 		// Fixes issue as result of https://github.com/easydigitaldownloads/easy-digital-downloads/issues/5490#issuecomment-283978899
 		if ( ! did_action( 'edd_purchase' ) ) {
@@ -123,8 +120,7 @@ class EDDUH_Track_History {
 		}
 
 		return $payment_meta;
-
-	} /* save_customer_history() */
+	}
 
 	/**
 	 * Handle developer debug data.
@@ -137,31 +133,27 @@ class EDDUH_Track_History {
 	 * @since 1.5.0
 	 */
 	public function devmode() {
-
 		// Only proceed if URL querystring cotnains "devmode=true"
-		if ( isset($_GET['devmode']) && 'true' == $_GET['devmode'] ) {
+		if ( defined( 'EDD_DEBUG' ) && EDD_DEBUG && isset( $_GET['devmode'] ) && 'true' == $_GET['devmode'] ) {
 
 			// Output user history if URL querystring contains 'output=history'
-			if ( isset($_GET['output']) && 'history' == $_GET['output'] ) {
+			if ( isset( $_GET['output'] ) && 'history' == $_GET['output'] ) {
 				var_dump( EDDUH_Cookie_Helper::get_cookie() );
 				echo '<pre>' . print_r( $this->get_customer_history(), 1 ) . '</pre>';
 			}
 
 			// Output user history cookie if URL querystring contains 'output=cookie'
-			if ( isset($_GET['output']) && 'cookie' == $_GET['output'] ) {
+			if ( isset( $_GET['output'] ) && 'cookie' == $_GET['output'] ) {
 				echo '<pre>' . print_r( $_COOKIE, 1 ) . '</pre>';
 			}
 
 			// Clear customer history and dump us back at the homepage if URL querystring contains 'history=reset'
-			if ( isset($_GET['history']) && 'reset' == $_GET['history'] ) {
+			if ( isset( $_GET['history'] ) && 'reset' == $_GET['history'] ) {
 				EDDUH_Cookie_Helper::delete_history_data();
 				wp_redirect( site_url() );
 				exit;
 			}
 
 		}
-
-	} /* devmode() */
-
+	}
 }
-return new edduh_Track_History;
